applied suggestions
