+++ CODEOWNERS
@@ -23,6 +23,7 @@
 CODEOWNERS @MrAlias @MadVikingGod @pellared @dashpole
 
 bridges/prometheus/                                                     @open-telemetry/go-approvers @dashpole
+bridges/otelzap/                                                        @open-telemetry/go-approvers @pellared @khushijain21
 
 config/                                                                 @open-telemetry/go-approvers @MadVikingGod @pellared @codeboten
